requirements.txt
fastapi
uvicorn
httpx
python-dotenv
asyncpg
google-generativeai
sqlalchemy[asyncio]
slack-sdk
bcrypt